+++ libs/expr/src/CRuntimeCompiledExpression_unittest.cpp
@@ -25,3 +25,32 @@ TEST(RuntimeCompiledExpression, SimpleTest)
 	EXPECT_NEAR(
 		expr.eval(), vars["x"] * vars["x"] + vars["x"] * vars["y"] + 1.0, 1e-9);
 }
+
+static double myNullary() { return 10.0; }
+
+static double myNeg(double x) { return -x; }
+
+static double myAdd(double x, double y) { return x + y; }
+
+static double myDiff(double x, double y, double z) { return (y - x) / z; }
+
+TEST(RuntimeCompiledExpression, CustomFunctions)
+{
+	mrpt::expr::CRuntimeCompiledExpression expr;
+	std::map<std::string, double> vars;
+
+	expr.register_function("myNullary", myNullary);
+	expr.register_function("myNeg", myNeg);
+	expr.register_function("myAdd", myAdd);
+	expr.register_function("myDiff", myDiff);
+
+	vars["x"] = 5.0;
+	vars["y"] = 3.0;
+	expr.compile("1+myAdd(x,y) + myNullary() + myNeg(x) + myDiff(x,y,x)", vars);
+
+	EXPECT_NEAR(
+		expr.eval(),
+		1 + vars["x"] + vars["y"] + 10.0 - vars["x"] +
+			(vars["y"] - vars["x"]) / vars["x"],
+		1e-9);
+}